+++ apps/docs/src/app/core/component-docs/date-picker/examples/date-picker-allow-null-example.component.ts
@@ -19,6 +19,6 @@ export class DatePickerAllowNullExampleComponent {
     date = FdDate.getNow();
 
     isInvalid(): boolean {
-        return !this.datePicker || !this.datePicker.isModelValid();
+        return !this.date || !this.date.isDateValid();
     }
 }